fix: Устранена циклическая зависимость импорта для стабильного запуска
Ранее возникала ошибка запуска сервиса (Gunicorn exit code 3), вызванная циклической зависимостью: `main.py` -> `api.py` -> `main.py`.

Это исправление решает проблему путем рефакторинга и вынесения общих компонентов (декораторов `login_required` и `setup_required`) в отдельный файл `app/decorators.py`.

Основные изменения:
- Создан новый файл `app/decorators.py`.
- `main.py` и `api.py` теперь импортируют декораторы из `app/decorators.py`, что разрывает цикл.

Это стандартная практика для Flask-приложений, которая обеспечивает корректную загрузку модулей и стабильный запуск приложения.
